Apply suggestions from code review
please work this time

+++ source/Preferences.h
@@ -30,7 +30,7 @@ public:
 		adaptive,
 	};
 
-	enum class DateFormat : int {
+	enum class DateFormat : int_fast8_t {
 		DMY = 0, // Day-first format. (Sat, 4 Oct 1941)
 		MDY,     // Month-first format. (Sat, Oct 4, 1941)
 		YMD      // All-numeric ISO 8601. (1941-10-04)
@@ -103,7 +103,7 @@ public:
 	static void ToggleAmmoUsage();
 	static std::string AmmoUsage();
 
-	// Date format preferences
+	// Date format preferences.
 	static void ToggleDateFormat();
 	static DateFormat GetDateFormat();
 	static const std::string &DateFormatSetting();
